fix(tabby-ui): rename route
