+++ ee/tabby-ui/app/files/components/file-tree.tsx
