+++ ee/tabby-ui/app/files/components/source-code-browser.tsx
