+++ ee/tabby-ui/app/files/components/source-code-editor.tsx
